Refactor user menu and API middleware; add fetchUser function for session handling
