+++ apps/app/app/lib/middlewares/auth-middleware.ts
