+++ apps/app/app/router.tsx
@@ -6,26 +6,26 @@ import { ErrorBoundary } from "~/components/error-boundary";
 import { routeTree } from "./routeTree.gen";
 
 export function createRouter() {
-	const queryClient = new QueryClient();
+  const queryClient = new QueryClient();
 
-	const router = createTanStackRouter({
-		routeTree,
-		context: { queryClient },
-		defaultPreload: "intent",
-		scrollRestoration: true,
-		defaultViewTransition: true,
-		defaultErrorComponent: ErrorBoundary,
-	});
+  const router = createTanStackRouter({
+    routeTree,
+    context: { queryClient },
+    defaultPreload: "intent",
+    scrollRestoration: true,
+    defaultViewTransition: true,
+    defaultErrorComponent: ErrorBoundary,
+  });
 
-	return routerWithQueryClient(router, queryClient);
+  return routerWithQueryClient(router, queryClient);
 }
 
 declare module "@tanstack/react-router" {
-	interface Register {
-		router: ReturnType<typeof createRouter>;
-	}
+  interface Register {
+    router: ReturnType<typeof createRouter>;
+  }
 
-	interface StaticDataRouteOption {
-		breadcrumb?: (data?: unknown) => React.ReactNode;
-	}
+  interface StaticDataRouteOption {
+    breadcrumb?: (data?: unknown) => React.ReactNode;
+  }
 }
